added json input format and converting with adapter

+++ FamilyRegistration.Core/FamilyRegistrationContext.cs
@@ -5,7 +5,7 @@ namespace FamilyRegistration.Core;
 public class FamilyRegistrationContext : Context
 {
     public required string Key { get; set; }
-    public int FamilyIncome { get; set; }
+    public decimal FamilyIncome { get; set; }
     public int NumOfDependents { get; set; }
     public int Score { get; private set; } = 0;
     public Dictionary<Type, int> Bag { get; } = new();
