+++ FamilyRegistration.Core/UseCases/ProcessData/InputItem.cs
@@ -3,7 +3,7 @@
 public class InputItem
 {
     public required string Key { get; set; }
-    public int FamilyIncome { get; set; }
+    public decimal FamilyIncome { get; set; }
     public int NumOfDependents { get; set; }
 
 }
