+++ FamilyRegistration.Core/UseCases/ProcessData/OutputItem.cs
@@ -3,7 +3,7 @@
 public class OutputItem
 {
     public required string Key { get; set; }
-    public int FamilyIncome { get; set; }
+    public decimal FamilyIncome { get; set; }
     public int NumOfDependents { get; set; }
 
     public required int Score { get; set; }
